+++ util.h
@@ -1,5 +1,5 @@
-int*** malloc_3dim_array(int, int, int);
-int** malloc_2dim_array(int, int);
-void free_3dim_array(int***, int, int);
-void free_2dim_array(int**, int);
-void print_2dim_array(int**, int, int);
+double*** malloc_3dim_array(int, int, int);
+double** malloc_2dim_array(int, int);
+void free_3dim_array(double***, int, int);
+void free_2dim_array(double**, int);
+void print_2dim_array(double**, int, int);
